feat:count record cases by AppId and list them
