+++ arex-storage-model/src/main/java/com/arextest/model/replay/CountRecordCaseResponseType.java
@@ -4,13 +4,9 @@ import com.arextest.model.response.Response;
 import com.arextest.model.response.ResponseStatusType;
 import lombok.Data;
 
-import java.util.List;
-
 @Data
 public class CountRecordCaseResponseType implements Response {
     private ResponseStatusType responseStatusType;
 
     private long recordedCaseCount;
-
-    private List<String> operationNameList;
 }
