+++ arex-storage-model/src/main/java/com/arextest/model/replay/ListRecordCaseRequestType.java
@@ -6,6 +6,7 @@ import lombok.Data;
 public class ListRecordCaseRequestType {
     private String appId;
     private String operationName;
+    private String operationType;
     private Integer pageSize;
     private String lastId;
 }
