+++ arex-storage-web-api/pom.xml
@@ -5,7 +5,7 @@
     <parent>
         <artifactId>arex-storage-service</artifactId>
         <groupId>com.arextest</groupId>
-        <version>1.0.4</version>
+        <version>1.0.5.2-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
     <artifactId>arex-storage-web-api</artifactId>
@@ -15,6 +15,7 @@
         <dependency>
             <groupId>com.arextest</groupId>
             <artifactId>arex-model</artifactId>
+            <version>1.0.5.2-SNAPSHOT</version>
         </dependency>
         <dependency>
             <groupId>com.arextest</groupId>
